backend/pyproject.toml
[project]
name = "gencrawl"
version = "1.0.0"
description = "General-purpose LLM-ready web crawler with natural language interface"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlmodel>=0.0.14",
    "psycopg2-binary>=2.9.9",
    "redis>=5.0.0",
    "celery>=5.3.4",
    "scrapy>=2.11.0",
    "playwright>=1.40.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "pymupdf>=1.23.0",
    "pdfplumber>=0.10.0",
    "pytesseract>=0.3.10",
    "pillow>=10.1.0",
    "psutil>=5.9.8",
    "anthropic>=0.8.0",
    "openai>=1.5.0",
    "weaviate-client>=3.26.0",
    "python-multipart>=0.0.6",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "jsonlines>=4.0.0",
    "apscheduler>=3.10.4",
    "croniter>=2.0.0",
    "nv-ingest-client==25.9.0",
    "nv-ingest-api==25.9.0",
]

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
